backend/requirements.txt
flask
flask-jwt-extended
rembg
Pillow
werkzeug
